Jetpack Cloud: move /manage/pricing redirect to the server
The Jetpack Manage pricing page has been disabled since #90190, with the
rendering controller commented out and the route redirecting to the
Jetpack.com For Agencies landing page. Because the section is isomorphic,
that redirect cost a full server render plus a client boot and a 1,536-line
chunk download before `window.location.assign` fired.

Move the redirect into a new `jetpackCloudPages()` in the express layer,
alongside `wpcomPages()`, and delete the client section. Query args are no
longer forwarded to the third-party landing page.
