--- client/jetpack-cloud/sections/manage/pricing/header/style.scss
@@ -1,52 +1,0 @@
-@import "@wordpress/base-styles/breakpoints";
-@import "@wordpress/base-styles/mixins";
-@import "@wordpress/base-styles/colors";
-
-.jetpack-product-store .header {
-	text-align: center;
-
-	.header__sub-title {
-		padding-top: 2em;
-		color: var(--studio-black);
-	}
-
-	.header__main-title {
-		h1 {
-			font-size: 2.5em;
-			font-weight: 600;
-
-			margin: 1.5em auto 0.25em;
-		}
-	}
-}
-
-.is-group-jetpack-cloud.is-section-jetpack-cloud-manage-pricing
-.header
-.header__main-title
-.formatted-header__title {
-	max-width: 280px;
-	margin-bottom: 0;
-	margin-right: auto;
-	margin-left: auto;
-
-	font-size: 1.5rem;
-	font-weight: 700;
-	letter-spacing: -1px;
-	line-height: 1.4;
-
-	@include break-small {
-		max-width: none;
-
-		font-size: 2rem;
-		line-height: 1.2;
-	}
-
-}
-
-.theme-jetpack-cloud .layout.is-section-jetpack-cloud-manage-pricing.has-no-masterbar {
-	padding-top: 0;
-}
-
-.is-group-jetpack-cloud.is-section-jetpack-cloud-manage-pricing div.jpcom-masterbar header.header {
-	margin-bottom: 0;
-}
