--- client/jetpack-cloud/sections/manage/pricing/package.json
@@ -1,4 +1,0 @@
-{
-	"main": "index.node.ts",
-	"browser": "index.web.js"
-}
