+++ client/my-sites/plans/jetpack-plans/product-store/style.scss
@@ -62,11 +62,6 @@
 .jetpack-product-store
 .header
 .header__main-title
-.formatted-header__title,
-.is-group-jetpack-cloud.is-section-jetpack-cloud-manage-pricing
-.jetpack-product-store
-.header
-.header__main-title
 .formatted-header__title {
 	margin: 0 auto;
 	font-size: $font-title-large;
@@ -176,9 +171,6 @@
 			text-decoration: none;
 
 			.is-group-jetpack-cloud.is-section-jetpack-cloud-pricing & {
-				background: var(--studio-gray-5);
-			}
-			.is-group-jetpack-cloud.is-section-jetpack-cloud-manage-pricing & {
 				background: var(--studio-gray-5);
 			}
 		}
