+++ client/sections.js
@@ -1,5 +1,3 @@
-const anyLocaleRegex = '([a-z]{2,3}|[a-z]{2}-[a-z]{2})';
-
 const sections = [
 	{
 		name: 'root',
@@ -715,14 +713,6 @@ const sections = [
 				href: 'https://fonts.googleapis.com/css2?family=Noto+Sans:ital,wght@0,400;0,700;1,400;1,700&display=swap',
 			},
 		],
-	},
-	{
-		name: 'jetpack-cloud-manage-pricing',
-		paths: [ '/manage/pricing', `/${ anyLocaleRegex }/manage/pricing` ],
-		module: 'calypso/jetpack-cloud/sections/manage/pricing',
-		group: 'jetpack-cloud',
-		enableLoggedOut: true,
-		isomorphic: true,
 	},
 	{
 		name: 'jetpack-cloud-features-comparison',
