+++ client/server/pages/index.js
@@ -8,6 +8,7 @@ import {
 	filterLanguageRevisions,
 	isTranslatedIncompletely,
 	isDefaultLocale,
+	getAnyLanguageRouteParam,
 	getLanguageSlugs,
 	localizeUrl,
 } from '@automattic/i18n-utils';
@@ -35,6 +36,7 @@ import {
 } from 'calypso/dashboard/app-dotcom/section';
 import { A4A_SIGNUP_PATHS } from 'calypso/dashboard/section';
 import isDashboardEnv from 'calypso/dashboard/utils/is-dashboard-env';
+import { JETPACK_COM_A4A_LANDING_PAGE } from 'calypso/jetpack-cloud/constants';
 import wooDnaConfig from 'calypso/jetpack-connect/woo-dna-config';
 import { STEPPER_SECTION_DEFINITION } from 'calypso/landing/stepper/section';
 import { SUBSCRIPTIONS_SECTION_DEFINITION } from 'calypso/landing/subscriptions/section';
@@ -1214,6 +1216,19 @@ function wpcomPages( app ) {
 	} );
 }
 
+function jetpackCloudPages( app ) {
+	const anyLangParam = getAnyLanguageRouteParam();
+
+	// The Jetpack Manage pricing page is disabled; send visitors to the Jetpack.com For
+	// Agencies landing page instead. A 302 rather than a 301 so the page can be restored
+	// without waiting out caches. The landing page is English-only, so no locale is carried
+	// over — see https://github.com/Automattic/wp-calypso/pull/90190.
+	// Query args are intentionally dropped rather than forwarded to a third-party domain.
+	app.get( [ '/manage/pricing', `/${ anyLangParam }/manage/pricing` ], function ( _req, res ) {
+		res.redirect( 302, JETPACK_COM_A4A_LANDING_PAGE );
+	} );
+}
+
 export default function pages() {
 	const app = express();
 
@@ -1228,6 +1243,11 @@ export default function pages() {
 
 	if ( ! ( isJetpackCloud() || isA8CForAgencies() || isDashboardEnv() ) ) {
 		wpcomPages( app );
+	}
+
+	// Registered before the section paths below, since express matches in registration order.
+	if ( isJetpackCloud() ) {
+		jetpackCloudPages( app );
 	}
 
 	/**
