+++ config/_shared.json
@@ -197,7 +197,6 @@
 		"jetpack-cloud-auth": false,
 		"jetpack-cloud-partner-portal": false,
 		"jetpack-cloud-pricing": false,
-		"jetpack-cloud-manage-pricing": false,
 		"jetpack-cloud-settings": false,
 		"jetpack-cloud-golden-token": false,
 		"jetpack-social": false,
