+++ config/jetpack-cloud-horizon.json
@@ -78,7 +78,6 @@
 		"jetpack-cloud-partner-portal": true,
 		"jetpack-cloud-plugin-management": true,
 		"jetpack-cloud-pricing": true,
-		"jetpack-cloud-manage-pricing": true,
 		"jetpack-cloud-settings": true,
 		"jetpack-cloud-golden-token": false,
 		"jetpack-search": true,
